+++ relevanceai/operations_new/processing/text/html_clean/ops.py
@@ -1,139 +1,12 @@
 """
 Clean HTML
 """
-import string
-import warnings
-from copy import deepcopy
-from relevanceai.operations_new.base import OperationBase
-from relevanceai.utils import MissingPackageError
-from typing import List, Optional
-from collections import Counter
-from html.parser import HTMLParser
-from io import StringIO
+from relevanceai.operations_new.processing.text.html_clean.base import (
+    CleanTextBase,
+)
 
 
-class BaseTextProcessing:
-    """Base text processing"""
-
-    @staticmethod
-    def normalize_text(
-        txt: str,
-        lower: bool = True,
-        remove_digit: bool = True,
-        remove_punct: bool = True,
-    ) -> str:
-        """
-        * Lower-casing
-        * Digit removal
-        * Punctuation removal
-        """
-        if lower:
-            txt = txt.lower()
-        if remove_digit:
-            txt = "".join([ch for ch in txt if ch not in string.digits])
-        if remove_punct:
-            txt = "".join([ch for ch in txt if ch not in string.punctuation])
-        return txt
-
-    @staticmethod
-    def get_word_frequency(
-        str_list: List[str],
-        remove_stop_words: bool = True,
-        additional_stop_words: Optional[List[str]] = None,
-        language="english",
-    ) -> List:
-        """Returns a sorted word frequency in Python"""
-        additional_stop_words = (
-            [] if additional_stop_words is None else additional_stop_words
-        )
-        try:
-            import nltk
-
-            nltk.download("stopwords")
-            from nltk.corpus import stopwords
-        except ModuleNotFoundError:
-            warnings.warn("You are missing NLTK, please run `pip install nltk`")
-
-        if remove_stop_words:
-            stpw = stopwords.words(language)
-            stpw += additional_stop_words
-        else:
-            stpw = []
-        word_counter = Counter(
-            [w.lower() for s in str_list for w in s.split() if w not in stpw]
-        )
-        return sorted(word_counter.items(), key=lambda item: (-item[1], item[0]))
-
-
-class MLStripper(HTMLParser):
-    """Remove HTML from the code and retrieves data."""
-
-    def __init__(self):
-        super().__init__()
-        self.reset()
-        self.strict = False
-        self.convert_charrefs = True
-        self.text = StringIO()
-
-    def handle_data(self, d):
-        self.text.write(d)
-
-    def get_data(self):
-        return (
-            self.text.getvalue()
-            .replace("\r", "")
-            .replace("\n", "")
-            .replace("\t", " ")
-            .strip()
-        )
-
-    def clean(self, text):
-        self.reset()
-        self.handle_data(text)
-        return self.get_data()
-
-
-class CleanTextOps(OperationBase):
+class CleanTextOps(CleanTextBase):
     """
     Clean text operations
     """
-
-    def clean_text(self, text):
-        """
-        Clean the text of the individuals
-        """
-        text = BaseTextProcessing.normalize_text(
-            text, lower=False, remove_punct=False, remove_digit=False
-        )
-        stripper = MLStripper()
-        return stripper.clean(text)
-
-    def clean_text_document(
-        self, text_field, document, output_field: str = "_cleantext_"
-    ):
-        """
-        Split a text field and store it in other values
-        """
-        t = self.get_field(text_field, document)
-        clean_text = self.clean_text(t)
-        # Format the split text into documents
-        self.set_field(output_field + "." + text_field, document, clean_text)
-        return document
-
-    def run(
-        self,
-        text_field,
-        documents,
-        inplace: bool = True,
-        output_field: str = "_cleantext_",
-    ):
-        if not inplace:
-            documents = deepcopy(documents)
-        # TODO; switch to something faster than list comprehension
-        [
-            self.clean_text_document(
-                text_field=text_field, document=document, output_field=output_field
-            )
-            for document in documents
-        ]
-        return documents
